finally a nice camera flight showing exit gate

+++ Dungeon-FrontEnd-TouchGUI/src/main/java/de/jdungeon/app/movieSequence/ChangeScaleSequence.java
@@ -1,6 +1,6 @@
 package de.jdungeon.app.movieSequence;
 
 public interface ChangeScaleSequence {
-	int getScale(float timePassed);
+	float getScale(float timePassed);
 
 }
